Fix MSVC compiler warning
reg.: '=': conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/importexport/videoexport/internal/videowriter.cpp
@@ -501,7 +501,7 @@ bool VideoWriter::generateScoreFrames(muse::media::IVideoEncoderPtr encoder, INo
         }
 
         INotationPainting::Options opt;
-        opt.fromPage = page->pageNumber();
+        opt.fromPage = static_cast<int>(page->pageNumber());
         opt.toPage = opt.fromPage;
         opt.deviceDpi = CANVAS_DPI;
